+++ spyrmsd/graphs/nx.py
@@ -4,6 +4,7 @@ from typing import Any, List, Optional, Tuple, Union
 import networkx as nx
 import numpy as np
 
+from spyrmsd.exceptions import NonIsomorphicGraphs
 from spyrmsd.graphs._common import (
     error_non_isomorphic_graphs,
     warn_disconnected_graph,
@@ -65,7 +66,7 @@ def match_graphs(G1, G2) -> List[Tuple[List[int], List[int]]]:
 
     Raises
     ------
-    ValueError
+    NonIsomorphicGraphs
         If the graphs `G1` and `G2` are not isomorphic
     """
 
@@ -92,8 +93,7 @@ def match_graphs(G1, G2) -> List[Tuple[List[int], List[int]]]:
 
     # Check if graphs are actually isomorphic
     if not GM.is_isomorphic():
-        # TODO: Create a new exception
-        raise ValueError(error_non_isomorphic_graphs)
+        raise NonIsomorphicGraphs(error_non_isomorphic_graphs)
 
     return [
         (list(isomorphism.keys()), list(isomorphism.values()))